Fix phantom error of missing testing directory
This error occcured during the invocation of go test -v ./... , seemingly randomly the testData
directory (where the cached results for the current test was saving to) would be deleted and the
build would fail as the program tried to save to a non existant folder. Even worse was this occured
consistently on github actions but rarely on my own machine.

To say this error has caused me grief would be an understatement. Test suites would run seemingly
perfectly on my own machine yet they'd fail consistently on the github actions environment. I tried
it all, I replicated the test environment to the best of my ability with cloning the repo on a
brand new directory and even machine to no avail. I put in dozens of print statements all over
the codebase hoping to find some rogue call to a function that was deleting the testData directory,
still nothing came up.

The only places in the codebase that were able to delete this directory were in the seperate setup
and teardown functions for each module's test suites. How could the folder be getting deleted
if the call to delete it had not even run yet? Well, as is common with concurrent programming you're
never sure what the other threads of execution are doing.

Behind the scenes go test -v ./... would run more than one module's tests at the same time. This
mean that as one shorter test suite finished it deleted the testData directory, which should be fine
as only that module is being tested. But this wrecked havoc as all modules use the same named directory
and this deleted it's own testData and the other modules directory. It took me a long time to even
think to consult google about go test running concurrently but in the end was a very quick fix.

I would like to specifically thank Lucas Roesler as his article was the search result that caught
my eye. https://lucasroesler.com/2019/02/when-sequential-tests-arent/